Remove v1 planning application creation service
