--- app/jobs/post_application_to_staging_job.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-class PostApplicationToStagingJob < ApplicationJob
-  queue_as :low_priority
-
-  def perform(local_authority, planning_application)
-    if (submission = planning_application.params_v2)
-      Apis::Bops::Query.new.post(local_authority.subdomain, submission)
-    else
-      Appsignal.send_error("Unable to find submission data for planning application with id: #{planning_application.id}")
-    end
-  end
-end
