+++ engines/bops_api/app/controllers/concerns/bops_api/schema_validation.rb
@@ -15,8 +15,7 @@ module BopsApi
           schema = Schemas.find!(name, schema: request_schema)
 
           unless schema.valid?(request_parameters)
-            raise BopsApi::Errors::InvalidRequestError,
-              "We couldn’t process your request because some information is missing or incorrect with error: #{schema.validate(request_parameters).to_a}"
+            raise BopsApi::Errors::InvalidRequestError, "We couldn’t process your request because some information is missing or incorrect."
           end
         end
       end
