+++ engines/bops_api/app/services/bops_api/application/creation_service.rb
@@ -79,6 +79,7 @@ module BopsApi
         end
 
         CreateNeighbourBoundaryGeojsonJob.perform_later(planning_application) if planning_application.consultation
+        PostApplicationToStagingJob.perform_later(local_authority, planning_application) if Bops.env.production?
 
         planning_application
       end
